refactor(client/redux, client/store): reestructuración, nuevo flujo de datos
