+++ client/src/redux/actions/actionFront.js
@@ -1,0 +1,1 @@
+import * as TYPES from '../types/index';
