+++ client/src/redux/actions/request.js
@@ -1,0 +1,9 @@
+import * as TYPES from '../types/index';
+
+//LOADING ACTIONS
+export const requestData = () => ({
+    type: TYPES.REQUEST_DATA,
+})
+export const requestSuccess = () => ({
+    type: TYPES.REQUEST_SUCCESS,
+})
