+++ client/src/redux/reducers/reducerTest.js
@@ -1,0 +1,24 @@
+// /*eslint-disable*/
+
+// import * as TYPES from '../types/index';
+
+// const initialState = {
+
+//   testCache: [],
+// };
+
+// export default function reducerTest (state = initialState, action){
+
+//   switch (action.type) {
+
+//     // PARA EJEMPLO
+//     //case TYPES.REQUEST_DATA: // Setea el LOADING en TRUE
+//     //   return {
+//     //     ...state,
+//     //     loading: true
+//     //   }
+
+//     default:
+//       return state;
+//   }
+// };
