+++ client/src/redux/types/index.js
@@ -1,13 +1,23 @@
 /*eslint-disable*/
+
+//PRODUCTS 
 export const GET_ALL_PRODUCTS = 'GET_ALL_PRODUCTS';
 export const GET_ONE_PRODUCT = 'GET_ONE_PRODUCT';
+
+//ARTIST
 export const FILTER_ARTISTS = 'FILTER_ARTISTS';
+export const SEARCH_BY_ARTIST = 'SEARCH_BY_ARTIST';
+
+//CATEGORIES
 export const NEW_CATEGORY = 'NEW_CATEGORY';
+export const GET_ALL_CATEGORIES = 'GET_ALL_CATEGORIES';
+
+//AUTHORS
 export const NEW_AUTHOR = 'NEW_AUTHOR';
 export const GET_ALL_AUTHORS = 'GET_ALL_AUTHORS';
-export const GET_ALL_CATEGORIES = 'GET_ALL_CATEGORIES';
+
+//SERIES
 export const GET_ALL_SERIES = 'GET_ALL_SERIES';
-export const SEARCH_BY_ARTIST = 'SEARCH_BY_ARTIST';
 
 //LOADING
 export const REQUEST_DATA = 'REQUEST_DATA'
