+++ client/src/store/index.js
@@ -1,11 +1,12 @@
-import { createStore, applyMiddleware } from 'redux';
+import { createStore, applyMiddleware, } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 
-import rootReducer from '../redux/reducers/index';
+import combineReducers from '../redux/reducers/combineReducers.js';
 
 const store = createStore(
-  rootReducer,
+
+  combineReducers,
   composeWithDevTools(applyMiddleware(thunk)),
 );
